requirements.txt
fastapi
uvicorn
python-multipart

python-dotenv
pydantic
pydantic-settings

openai
litellm
crewai
instructor
tenacity

numpy


pymongo


redis
upstash-redis

pandas
openpyxl
pdfplumber

requests
tqdm
orjson
rich
